ansible: set hostname to actual server hostname
This will allow to immediately know which server is actually producing
error messages, for example.

SQUAD_HOSTNAME is used to compose SQUAD_BASE_URL is the later is not
specified, but otherwise is not used. By setting SQUAD_BASE_URL
explicitly, we are free to use SQUAD_HOSTNAME in any way that we want.