k8s: fetch- worker: make squad-worker the container main process
The gracePeriod implemented a few commits back weren't working as
excpected because the container was being started as:

  sh -c "squad-worker ..."

Therefore, when K8S sent SIGKILL to the container, only `sh` would die,
and yet the worker would still be running normally until the end of
gracePeriod, when it would be force-killed and lose progress.

This patch also adds the `--without-gossip`, which reduces communication
among workers and reduces probability of memory leaks.