terraform: s3: enable S3 Glacier backups
This will replicate any incoming data to S3 bucket into
a backup bucket with "Glacier Deep Archive" storage type [1].

This storage type costs $ 0.00099 / GB / month. It's dirt cheap
and it's made for cases where rare access is required, like
once or twice a year.

Only drawback of this storage type is the retriving time. It might
take up to 12 hours to retrive a file storage in Deep Archive mode.

Storing files in standard S3 storage already has a redundancy, so having
a backup is a fall back of a fall back if data ever gets lost at some
point.

[1] https://aws.amazon.com/blogs/aws/new-amazon-s3-storage-class-glacier-deep-archive/