k8s: qareports-extra-settings: enable health-check
The idea of extra settings is to have a file that gets
evaluated right in the end of squad.settings.py file.

The extra settings contains additional django-health-check
imports.

Keep in mind that it always try to load ldap config, but that one
is available only in the frontend pod.

Signed-off-by: Charles Oliveira <charles.oliveira@linaro.org>